+++ src/buffer.rs
@@ -156,10 +156,12 @@ pub unsafe trait Element: Copy {
 fn validate(b: &ffi::Py_buffer) -> PyResult<()> {
     // shape and stride information must be provided when we use PyBUF_FULL_RO
     if b.shape.is_null() {
-        return Err(exceptions::BufferError::py_err("Shape is Null"));
+        return Err(exceptions::PyBufferError::py_err("Shape is Null"));
     }
     if b.strides.is_null() {
-        return Err(exceptions::BufferError::py_err("PyBuffer: Strides is Null"));
+        return Err(exceptions::PyBufferError::py_err(
+            "PyBuffer: Strides is Null",
+        ));
     }
     Ok(())
 }
@@ -188,7 +190,7 @@ impl<T: Element> PyBuffer<T> {
             {
                 Ok(buf)
             } else {
-                Err(exceptions::BufferError::py_err(
+                Err(exceptions::PyBufferError::py_err(
                     "Incompatible type as buffer",
                 ))
             }
@@ -439,7 +441,7 @@ impl<T: Element> PyBuffer<T> {
 
     fn copy_to_slice_impl(&self, py: Python, target: &mut [T], fort: u8) -> PyResult<()> {
         if mem::size_of_val(target) != self.len_bytes() {
-            return Err(exceptions::BufferError::py_err(
+            return Err(exceptions::PyBufferError::py_err(
                 "Slice length does not match buffer length.",
             ));
         }
@@ -526,7 +528,7 @@ impl<T: Element> PyBuffer<T> {
             return buffer_readonly_error();
         }
         if mem::size_of_val(source) != self.len_bytes() {
-            return Err(exceptions::BufferError::py_err(
+            return Err(exceptions::PyBufferError::py_err(
                 "Slice length does not match buffer length.",
             ));
         }
@@ -562,7 +564,7 @@ impl<T: Element> PyBuffer<T> {
 
 #[inline(always)]
 fn buffer_readonly_error() -> PyResult<()> {
-    Err(exceptions::BufferError::py_err(
+    Err(exceptions::PyBufferError::py_err(
         "Cannot write to read-only buffer.",
     ))
 }