+++ src/panic.rs
@@ -1,12 +1,13 @@
-use crate::exceptions::BaseException;
+use crate::exceptions::PyBaseException;
 
-/// The exception raised when Rust code called from Python panics.
-///
-/// Like SystemExit, this exception is derived from BaseException so that
-/// it will typically propagate all the way through the stack and cause the
-/// Python interpreter to exit.
-pub struct PanicException {
-    _private: (),
-}
+pyo3_exception!(
+    "
+    The exception raised when Rust code called from Python panics.
 
-pyo3_exception!(PanicException, BaseException);
+    Like SystemExit, this exception is derived from BaseException so that
+    it will typically propagate all the way through the stack and cause the
+    Python interpreter to exit.
+    ",
+    PanicException,
+    PyBaseException
+);